Master

+++ lib/cocos2d-x/cocos2dx/cocoa/CCDictionary.cpp
@@ -107,7 +107,7 @@ CCArray* CCDictionary::allKeys()
     {
         HASH_ITER(hh, m_pElements, pElement, tmp) 
         {
-            CCInteger* pOneKey = new CCInteger(pElement->m_iKey);
+            CCInteger* pOneKey = new CCInteger((int)pElement->m_iKey);
             pArray->addObject(pOneKey);
             CC_SAFE_RELEASE(pOneKey);
         }
@@ -142,7 +142,7 @@ CCArray* CCDictionary::allKeysForObject(CCObject* object)
         {
             if (object == pElement->m_pObject)
             {
-                CCInteger* pOneKey = new CCInteger(pElement->m_iKey);
+                CCInteger* pOneKey = new CCInteger((int)pElement->m_iKey);
                 pArray->addObject(pOneKey);
                 CC_SAFE_RELEASE(pOneKey);
             }
