+++ lib/cocos2d-x/cocos2dx/cocoa/CCString.cpp
@@ -121,7 +121,7 @@ const char* CCString::getCString() const
 
 unsigned int CCString::length() const
 {
-    return m_sString.length();
+    return (unsigned int)m_sString.length();
 }
 
 int CCString::compare(const char * pStr) const
