+++ lib/cocos2d-x/cocos2dx/label_nodes/CCLabelAtlas.cpp
@@ -64,7 +64,7 @@ bool CCLabelAtlas::initWithString(const char *string, const char *charMapFile, u
 bool CCLabelAtlas::initWithString(const char *string, CCTexture2D* texture, unsigned int itemWidth, unsigned int itemHeight, unsigned int startCharMap)
 {
     CCAssert(string != NULL, "");
-    if (CCAtlasNode::initWithTexture(texture, itemWidth, itemHeight, strlen(string)))
+    if (CCAtlasNode::initWithTexture(texture, itemWidth, itemHeight, (unsigned int)strlen(string)))
     {
         m_uMapStartChar = startCharMap;
         this->setString(string);
@@ -114,7 +114,7 @@ bool CCLabelAtlas::initWithString(const char *theString, const char *fntFile)
 //CCLabelAtlas - Atlas generation
 void CCLabelAtlas::updateAtlasValues()
 {
-    unsigned int n = m_sString.length();
+    unsigned int n = (unsigned int)m_sString.length();
 
     const unsigned char *s = (unsigned char*)m_sString.c_str();
 
@@ -189,7 +189,7 @@ void CCLabelAtlas::updateAtlasValues()
 //CCLabelAtlas - CCLabelProtocol
 void CCLabelAtlas::setString(const char *label)
 {
-    unsigned int len = strlen(label);
+    unsigned int len = (unsigned int)strlen(label);
     if (len > m_pTextureAtlas->getTotalQuads())
     {
         m_pTextureAtlas->resizeCapacity(len);
